[FEAT] Adicionado relacionamento entre user e produto

+++ src/app/usecases/Product/CrudProductAction.ts
@@ -4,6 +4,7 @@ interface Product {
     description:      string;
     price:     number;
     type:      string;
+    userId: number;
 }
 class CrudProductAction{
 
@@ -23,14 +24,24 @@ class CrudProductAction{
         return {message: "Erro: Não foi possível encontrar o Produto!"};
     }
     public async updateProduct(id:number, product: Product){
-        const response = await productDao.updateProduct(id, product);
+        const productOld = await productDao.getProductById(id);
+        let response = null;
+        if(productOld && productOld.userId == product.userId){
+            response = await productDao.updateProduct(id, product);
+        }else{
+            return {message: "Erro: Você não tem permissão para isso !"};
+        }
         if (response) {
             return {message: "Produto atualizado com sucesso!"};
         }
         return {message: "Erro: Não foi possível atualizar o Produto!"};
     }   
-    public async deleteProduct(id:number){
-        const response = await productDao.deleteProduct(id);
+    public async deleteProduct(id:number, product: Product){
+        const productOld = await productDao.getProductById(id);
+        let response = null;
+        if(productOld && productOld.userId == product.userId){
+            response = await productDao.deleteProduct(id);
+        }
         if (response) {
             return {message: "Produto deletado com sucesso!"};
         }
